Add an option to automatically delete grib files older than 3 days
git-svn-id: svn://svn.code.sf.net/p/qtvlm/code/branches/qtVlm_dev@1772 2387b160-4621-4e8a-be36-40ea10d72da2

+++ src/mycentralwidget.cpp
@@ -1244,7 +1244,7 @@ void myCentralWidget::loadGribFile(QString fileName, bool zoom)
         QFileInfoList list=inf.absoluteDir().entryInfoList(f);
         foreach(const QFileInfo &i,list)
         {
-            if(i.lastModified()<QDateTime::currentDateTime().addDays(-3))
+            if(i.lastModified()<QDateTime::currentDateTime().addDays(-3) && i.filePath()!=inf.filePath())
                 QFile::remove(i.filePath());
         }
     }
